ML/for_cloud_function/requirements.txt
tensorflow==2.5.0
PySastrawi==1.2.0